Adapt peano.c for the new data structures

+++ src/proto.h
@@ -21,7 +21,7 @@
 #include "config.h"			// holds Config #defines
 
 /*
- * Other Code Parameters
+ * Unexposed Code Parameters
  */
 
 #define CHARBUFSIZE 256L	// Maximum No. of bytes in every char buffer
@@ -44,7 +44,7 @@ typedef uint32_t ID_t;
 #endif
 
 typedef uint32_t intime_t;		// type of integer time 
-typedef __uint128_t peanoKey; 	// long peanokey, 42 triplets / levels
+typedef uint64_t peanoKey; 	// long peanokey, 42 triplets / levels
 typedef uint64_t shortKey;		// short peanokey, 21 triplets / levels
 
 /* 
@@ -55,13 +55,13 @@ extern void Finish();
 extern void Print_Compile_Time_Settings();
 
 /* 
- * Workarounds 
+ * Workaround
  */
 
 double erand48(unsigned short xsubi[3]);
 
 /* 
- * global module headers  
+ * Global module headers  
  */
 
 #include "macro.h"			// macro definitions
@@ -82,7 +82,6 @@ double erand48(unsigned short xsubi[3]);
 #include "cosmology.h"		// cosmology functions 
 #include "comov.h"			// Comoving coordinates
 #include "periodic.h"		// periodic boundary conditions
-
 
 
 #endif // PROTO_H
